+++ src/commands/supervivence/general/sinv.js
@@ -1,0 +1,12 @@
+module.exports = ({
+    name: "sinv",
+    code: `$title[Inventario de Supervivencia]$description[
+        Personaje: $getUserVar[character]
+        Ficha de naturaleza: $getUserVar[naturetoken]
+        Ficha gamer: $getUserVar[gamertoken]
+        Jefes Derrotados: $getUserVar[defeatedbosses]
+        Vida del jugador: $getUserVar[characterlife]
+    ]$color[BLUE]
+    $onlyIf[$getGlobalUserVar[bl]==false;{title:Oh vaya! Eres malicioso...}{description:
+        :x:| Tu acceso al bot Akeno a sido denegado, parece que haz hecho algo :/, si crees que esto es un error, contacta el soporte y te ayudaremos}{color:RED}]`
+})